refactor: having null data on many properties in Transfer is valid.  It'll be often that some data is missing, and therefore some models will be null.  (And having the required properties helps from a code and intention perspective)
We don't want errors like JSON deserialization for type 'SUI.Transfer.Domain.ConformedData' was missing required properties including: 'childServicesReferralSummaries', 'crimeMissingEpisodesSummaries'. causing Transfer or View to fall over.  Its needs to be more graceful.

In Transfer (and View), we should definitely be of the mindset of graceful degradation, rather than erroring.  If a few fields are missing, there's still loads of value to displaying all the other data fields to the safeguarding users.