feat: SUI-1236 - consolidation of Custodian data (#281)
Interim pull request for SUI-1236 "Consolidation of custodian data".

Consolidates each field in every record type, for every record returned from the providers.

Order of precedence is:
1. Field ranking, so that we can ultimately rank certain providers above others, on a field by field basis
    - This PR only contains skeleton logic for this.  This is to be implemented in: https://hippodigital-dfe.atlassian.net/browse/SUI-1273
2. Provider ID as a fallback
3. Null values go to the back

Uses C# source generators to implement the logic, so that we’re not using reflection.

This PR relates to: https://hippodigital-dfe.atlassian.net/browse/SUI-1236

Follow on PR will integrate this code in to the actual Transfer process.